// src/app/page/master/master-pegawai-sk-hitung-payroll/master-pegawai-sk-hitung-payroll.interface.ts
import {EventEmitter} from '@angular/core';

export interface MasterPegawaiSkHitungPayroll {
	
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
	id?: any;
	kode?: any;
	version?: any;
	label?: any;
	code?: any;
	nama?: any;
	kodeEx?: any;
	namaEx?: any;
	value?: any;
 }